+++ packages/nimble-components/src/rich-text-viewer/tests/rich-text-viewer.spec.ts
@@ -1,0 +1,13 @@
+import { RichTextViewer, richTextViewerTag } from '..';
+
+describe('RichTextViewer', () => {
+    it('can construct an element instance', () => {
+        expect(
+            document.createElement('nimble-rich-text-viewer')
+        ).toBeInstanceOf(RichTextViewer);
+    });
+
+    it('should export its tag', () => {
+        expect(richTextViewerTag).toBe('nimble-rich-text-viewer');
+    });
+});