+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Tue, 27 Jun 2023 19:05:52 GMT and should not be manually modified.
+This log was last generated on Wed, 28 Jun 2023 13:37:50 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 16.6.10
+
+Wed, 28 Jun 2023 13:37:50 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v19.4.2
 
 ## 16.6.9